Renaming and documentation
