+++ api/management/commands/setup_test_delegates.py
@@ -1,8 +1,8 @@
 from django.core.management.base import BaseCommand
-from testdata.generate_testdata import generate_delegates
+from test_data.generate_test_data import generate_delegates
 
 class Command(BaseCommand):
-    help = 'Add test delegate to the database'
+    help = 'Add test delegates to the database'
 
     def add_arguments(self, parser):
         # Optional argument
