+++ api/management/commands/setup_test_mun_directors.py
@@ -1,8 +1,8 @@
 from django.core.management.base import BaseCommand
-from testdata.generate_testdata import generate_mun_directors
+from test_data.generate_test_data import generate_mun_directors
 
 class Command(BaseCommand):
-    help = 'Add test MUN director to the database'
+    help = 'Add test MUN directors to the database'
 
     def add_arguments(self, parser):
         # Optional argument
